refactor: move bookmark classes to new bookmark module and update package structure
